Add checks that labware is usable (not released or destroyed)
Check in plan validation that the source is usable (and dest does not
exist yet).
Check in confirmation that the destination is (still) usable.

At the moment I'm not rechecking the source labware at confirm because
if it *is* (for instance) released, that means it was released between
planning an operation and confirming it, and the only outcome of the
software blocking the late confirmation is that we get a patch request
to let them do what they're trying to do here.

Also, added more boolean state fields to labware type in graphql.