Check for tissue external name clashes in block registration
If there is a clash, return a result containing info about clashes.
If the request specifies that the clash is intentional, validate the
info given and then create new labware linked to existing tissue.

Remove the "tissues" field from RegisterResult, because now a successful
register does not necessarily create new tissues.

Remove logic that incorporates *highest section number from this tissue*
when calculating a new section number, because having multiple blocks
linked to the same tissue is now supported.