Printers can now have multiple label types
That gives us the option of using label type to identify a particular
layout and set of content for a particular purpose.
I added a new liquibase custom sql writer to set up printers with
their label types.
I added a created field into the labware entity class (corresponding
to an existing column in the table). Now the set of information that
can be put on a labware label includes the labware's creation date
(iso format).

Plates are going to use a tiny label for now, so the label type for them
is called "plate" rather than "thin". The "plate" template is copied
from the "tiny" template but with a bunch of fields removed and a date
field added.