x653 record result following extract op
A result specifies a pass with a concentration, or a fail with a reason,
for each specified item of labware. The concentration will be recorded
as a measurement. The reason is recorded as a comment. The pass/fail is
recorded as a result (ResultOp).

New comment category "extract result" set up in static data patch.