x766 support getting works along with the status comment (as a string)
Provide a way to get a list of works along with the comment
that was recorded when they entered their current state.
Comments are only recorded when a work is failed or paused.
Theoretically, if you are looking up a paused work, then its
latest event should be of type "pause" and there should be
a comment on it.