x753 extensive rewrite
Previously we would look up the storage locations in storelight
when a release file is requested. The problem with that is that
the labware typically are not stored anywhere any more because they
have already been released.
So now, we validate the request and look up storage locations before
the release transactions, and we save the storage information in
the labware_release record. The release file service then only
has to pull out the recorded storage information from the release table
instead of making a storelight query.

You'll have to drop your old unit test database and allow
liquibase to recreate it, because the schema has changed
to support storage storage locations in the labware_release table.