Fix following x838: make ancestoriser work with in-place sample changes
The dual index op introduced the possibility for an operation to
change the bio state of samples in existing labware. That means
it replaces the samples currently in a slot with new samples
with a different bio state. The ancestoriser assumed that any
action in-place should be ignored, so it failed to track ancestors
beyond this point. This prevented ancestral measurements being looked
up.