x858 code review
Use maps and sets more effectively (looking up in them rather than
iterating them).
op.getActions() isn't allowed to be null in prod code: if it is null,
that is a test problem.
Loading releases inside the op loop is unnecessary work: load them
after the loop.
Check for releases on any destination labware for the work: do not
assume that the labware should have "isReleased()". If labware is
unreleased, it will not have "isReleased()" any more.

Updated tests, including integration test.