x672: add Tissue.collectionDate
Support a "sampleCollectionDate" field in block registration.
The date is compulsory when registering human fetal samples; optional
for other samples. (Tissues already created can still have null for
the date.)
When existing tissue is registered, the date must be consistent;
except you may specify a date for tissue that previously had no date.
In that case the existing tissue will be updated with the given date.