FileStoreService guard against dupes
Make sure we don't overwrite existing files if we
somehow generate a path that is already in use.
Finish the StanFileRepo transaction before calling
transferTo: if there is an identical path already in
the table, the unique key will fail the transaction.

If the file path transfer itself errors, log the error
and delete the created StanFile.