Another go at preempting race conditions in file store service
Previously, if the file did fail to transfer, then the new file created
would be deleted; but the files that were deprecated would not be
undeprecated. Trying to manage that outside of a transaction is more
hassle.
New approach:
1: check if a file with the intended path exists in the database:
   if it does, throw an exception
2: transfer the file
3. Transaction: save the StanFile entity and deprecate others

All three steps are in a sync-block to obviate the possibility
that two similarly named files could be created simultaneously.

Added StanFileRepo.existsByPath