x989 Slot regions
A slot region is a name to describe a region in a labware slot.
A SamplePosition is the region a sample occupies in a particular slot.
Slot regions may be added and en-/disabled like other adminnable
elements.
SamplePositions may be recorded when confirming sectioning,
and queried via the new "samplePositions(labwareBarcode)" query.

Also, comments may now be recorded against specific samples in
a slot when confirming sections.