x1167 Make lifestage optional
The LifeStage field is now nullable in graphql and in the database.
The client can offer an n/a option for life stage as long as it
uses null in the request.
(Making n/a an acceptable value in the LifeStage enum would have
needed more extensive rewrite because a graphql enum cannot
contain a slash.)