From RT#801885 Fix possible transaction race condition
What seems to have happened is that two requests were executed
at nearly the same time (in separate transactions) and they
tried to set the same work's operation-ids. The second overwrote
the changes in the first. I believe that the way the code was written
(in particular, loading the contents of the field into a new arraylist,
modifying it, and then setting it back in the object) allowed this
to happen (though I don't think it should).

To avoid this, don't read the list contents into a new list and then
set it back; update it in place. To avoid dupes, change it to
a set instead of a list. It makes more sense as a set anyway.
Avoid using immutable collection objects so that we can always mutate
the collections instead of overwriting them.