RT#804644: Use extra decoding step for work numbers in file controller
Due to some quirk between the client and the core, the work numbers
in the URL parameter for storing a file are received in a URL-
encoded state, which makes them not valid work numbers.
Go through a step of decoding them to make sure that they are correctly
formed.